+++ src/common/display.ts
@@ -339,7 +339,7 @@ export class Display extends EventEmitter {
             this._el.removeChild(this.split);
             this.split = null;
         }
-        this.doUpdate(UpdateType.scroll | UpdateType.view | UpdateType.scrollView);
+        this.doUpdate(UpdateType.scrollbars | UpdateType.scroll | UpdateType.view | UpdateType.scrollView);
     }
 
     get linkFunction(): string {
@@ -882,7 +882,7 @@ export class Display extends EventEmitter {
             }
             if ((this._updating & UpdateType.scroll) === UpdateType.scroll) {
                 if (this.split) {
-                    if (this._VScroll.position >= this._VScroll.scrollSize - this._padding[0]) {
+                    if (this._VScroll.atBottom) {
                         this.split.style.display = '';
                         this.split.shown = false;
                         if (this._scrollCorner) this._scrollCorner.classList.remove('active');
@@ -2717,7 +2717,7 @@ export class Display extends EventEmitter {
         else
             this._VScroll.padding = 0;
 
-        if (!this._HScroll.visible && this._scrollCorner && !this.split && !this._showSplitButton) {
+        if (!this._HScroll.visible && this._scrollCorner && (!this.split || !this._showSplitButton)) {
             this._el.removeChild(this._scrollCorner);
             this._scrollCorner = null;
         }
@@ -2880,6 +2880,9 @@ export class ScrollBar extends EventEmitter {
         position: 0
     };
 
+    private $resizeObserver;
+    private $resizeObserverCache;
+
     get size(): number { return this._visible ? 12 : 0; }
 
     get position(): number { return this._position - (this._type === ScrollType.horizontal ? this._padding[3] : this._padding[0]); }
@@ -2889,6 +2892,7 @@ export class ScrollBar extends EventEmitter {
         if (value !== this._offset) {
             this._offset = value;
             this.updateLocation();
+            this.resize();
         }
     }
     get padding(): number { return this.$padding; }
@@ -2896,6 +2900,7 @@ export class ScrollBar extends EventEmitter {
         if (value !== this.$padding) {
             this.$padding = value;
             this.updateLocation();
+            this.resize();
         }
     }
 
@@ -2924,6 +2929,8 @@ export class ScrollBar extends EventEmitter {
             this.resize();
         }
     }
+
+    get atBottom(): boolean { return this.position >= this.scrollSize - this._padding[0]; }
 
     constructor(parent?: HTMLElement, content?: HTMLElement, type?: ScrollType) {
         super();
@@ -2971,6 +2978,7 @@ export class ScrollBar extends EventEmitter {
         this.track = document.createElement('div');
         this.track.className = 'scroll-track scroll-' + (this._type === ScrollType.horizontal ? 'horizontal' : 'vertical');
         this.track.style.position = 'absolute';
+        this.track.style.overflow = 'hidden';
         this.track.addEventListener('mousedown', (e) => {
             if (e.button === 0 && e.buttons) {
                 this._lastMouse = e;
@@ -3025,6 +3033,16 @@ export class ScrollBar extends EventEmitter {
 
         window.addEventListener('resize', this._wResize.bind(this));
         this.resize();
+        this.$resizeObserver = new ResizeObserver((entries, observer) => {
+            if (entries.length === 0) return;
+            if (!entries[0].contentRect || entries[0].contentRect.width === 0 || entries[0].contentRect.height === 0)
+                return;
+            if (!this.$resizeObserverCache || this.$resizeObserverCache.width !== entries[0].contentRect.width || this.$resizeObserverCache.height !== entries[0].contentRect.height) {
+                this.$resizeObserverCache = { width: entries[0].contentRect.width, height: entries[0].contentRect.height };
+                this.resize();
+            }
+        });
+        this.$resizeObserver.observe(this.track);
     }
 
     public reset() {
@@ -3043,9 +3061,9 @@ export class ScrollBar extends EventEmitter {
             this.track.classList.remove('scroll-disabled');
         else
             this.track.classList.add('scroll-disabled');
-        this._thumbSize = Math.ceil(1 / this._percentView * this._parentSize);
-        if (this._thumbSize > this._parentSize)
-            this._thumbSize = this._parentSize;
+        this._thumbSize = Math.ceil(1 / this._percentView * this._trackSize);
+        if (this._thumbSize > this._trackSize)
+            this._thumbSize = this._trackSize;
         if (this._thumbSize < 20)
             this._thumbSize = 20;
         this.thumb.style[this._type === ScrollType.horizontal ? 'width' : 'height'] = this._thumbSize + 'px';
@@ -3082,7 +3100,7 @@ export class ScrollBar extends EventEmitter {
         this.updatePosition(0);
     }
 
-    public resize(scrollToEnd?) {
+    public resize() {
         const pc = window.getComputedStyle(this._parent);
         this._padding = [
             parseInt(pc.getPropertyValue('padding-top')) || 0,
@@ -3090,7 +3108,7 @@ export class ScrollBar extends EventEmitter {
             parseInt(pc.getPropertyValue('padding-bottom')) || 0,
             parseInt(pc.getPropertyValue('padding-left')) || 0
         ];
-        const bottom = this.position >= this.scrollSize - this._padding[0];
+        const bottom = this.atBottom;
         if (this._type === ScrollType.horizontal) {
             this._contentSize = this._content.clientWidth + this._padding[1] + this._padding[3];
             this._parentSize = this._parent.clientWidth - this.offset - this._scrollOffset;
@@ -3114,7 +3132,9 @@ export class ScrollBar extends EventEmitter {
     }
 
     public currentPosition() {
-        const p = this._type === ScrollType.horizontal ? (this._lastMouse.pageX - this.state.position - this._os.left) : (this._lastMouse.pageY - this.state.position - this._os.top);
+        let p = this._type === ScrollType.horizontal ? (this._lastMouse.pageX - this.state.position - this._os.left) : (this._lastMouse.pageY - this.state.position - this._os.top);
+        if (p >= this._maxDrag)
+            p = this._maxDrag;
         if (p < 0)
             return 0;
         if (p > this.maxPosition)